Make project activation faster.
Project activation should be as fast as the auto complete for project
names. We should just use the same method for getting project names as
we do for finding the settings file and activating it.

Idealy. This should be moved to use a class so we can make the django
project a object instead of variouse autoload functions.