Create a new way to deal with project searchs.
We need a better way to search through projects and better explain what
it does. This commit removes the django_project_container and creates a
new vairable called django_rpoejct_search_exp. Using this we can make
more complex searches rather than before.

This also sets a default of "**" so it should be able to search through
all settings folders and files. This should fix #5